wip: moved logic to new filter class TeamProjectBasedAuthorizingFilter
